Features/combat pve

+++ infrastructure/src/main/java/com/arcathoria/monster/dto/FileMonsterDTO.java
@@ -2,6 +2,6 @@ package com.arcathoria.monster.dto;
 
 import java.util.UUID;
 
-public record FileMonsterDTO(UUID monsterId, String monsterName, double currentHealth, double maxHealth,
+public record FileMonsterDTO(UUID monsterId, String monsterName, int currentHealth, int maxHealth,
                              int intelligence) {
 }
